mips: make loongsoon serial driver always builtin
Paul noted this file looks as if it is non-modular, but it piggy-backs
off CONFIG_SERIAL_8250 which is tristate.  If set to "=m"
we will get this after the init/module header cleanup:

 arch/mips/loongson/common/serial.c:76:1: error: data definition has no type or storage class [-Werror]
 arch/mips/loongson/common/serial.c:76:1: error: type defaults to 'int' in declaration of 'device_initcall' [-Werror=implicit-int]
 arch/mips/loongson/common/serial.c:76:1: error: parameter names (without types) in function declaration [-Werror]
 arch/mips/loongson/common/serial.c:58:19: error: 'serial_init' defined but not used [-Werror=unused-function]
 cc1: all warnings being treated as errors
 make[3]: *** [arch/mips/loongson/common/serial.o] Error 1

Paul originally suggested making it explicitly modular, but Ralf noted that
such a commit would result in platform device registrations from module code
which opens another can of worms.  This and the whole philosophy of platforms
devices to show what devices do exist in a system, not which drivers are
configured.  Hence this commit just always build serial.c into the kernel.

A related issue is uart_base.o which I think is required to register
properly initialized platform devices.  It depends on
CONFIG_LOONGSON_UART_BASE but probably should also be put into the
kernel whenever we register the UART platform devices and that's
always.

Reported-by: Paul Gortmaker <paul.gortmaker@windriver.com>
Signed-off-by: Ralf Baechle <ralf@linux-mips.org>
[PG: tweaked commit log from e-mail exchange; see:
 https://lkml.kernel.org/r/20150609073533.GB2753@linux-mips.org ]
Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>