watchdog: avoid changing cpumask on unregistered smpboot thread
Also, add a line documenting that you are not supposed to try to do this.

The risk, as Ulrich Obergfell pointed out, is that if registration fails
due to ENOMEM, there may be no valid cpumask field in the smpboot object,
so attempting to modify it in smpboot_update_cpumask_percpu_thread() would
crash.

In the case of the watchdog this change is fine because we will reset the
smpboot cpumask as soon as we successfully manage to register the smpboot
thread.

Signed-off-by: Chris Metcalf <cmetcalf@ezchip.com>
Cc: Ulrich Obergfell <uobergfe@redhat.com>
Cc: Don Zickus <dzickus@redhat.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>