vfs: Commit to never having exectuables on proc and sysfs.
Today proc and sysfs do not contain any executable files.  Several
applications today mount proc or sysfs without noexec and nosuid
in situations where it would be a security issue if there were
actually exectuables files on proc or sysfs.

Therefore commit to never allowing executables on proc and sysfs
by adding a new flag, to mark them as filesystems without executables
and enforce that flag.

Test the flag where MNT_NOEXEC is tested today, so that the only
user visible effect will be that exectuables will be treated
as if the execute bit is cleared.

The filesystems proc and sysfs do not currently incoporate any
executable files so this does not result in any user visible effects.

This makes it unnecessary to vet changes to proc and sysfs tightly for
adding exectuable files or changes to chattr that would modify
existing files, as no matter what the individual file say they will
not be treated as exectuable files by the vfs.

Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>