ARM: shmobile: r8a7779 Marzen legacy fix V2
Unbreak the r8a7779 Marzen legacy code provided by marzen_defconfig.

As it is today Marzen multiplatform is working, but the legacy code
is broken. This patch intends to leave multiplatform as-is but do
a simple one-shot fix to unbreak the legacy code base.

Without this patch there is no preset delay and the TWD is defined
both in C code and in DT. Solve the delay by invoking shmobile_init_delay()
and simply disable TWD in case of building for legacy.

Marzen legacy code will be removed in the near future, but until then
we may as well avoid breaking it.

Signed-off-by: Magnus Damm <damm+renesas@opensource.se>
Fixes: f2a5473861cf ("clocksource: sh_tmu: Set cpu_possible_mask to fix SMP broadcast")
Tested-by: Simon Horman <horms+renesas@verge.net.au>
Signed-off-by: Simon Horman <horms+renesas@verge.net.au>