nfsd: convert laundry_wq to something less nfsd4 specific
Currently, nfsd uses a singlethread workqueue for this, but that's
probably not necessary. If we have multiple namespaces, then there's no
need to serialize the laundromat runs. They can run in parallel.

Also, create_singlethread_workqueue adds the WQ_MEM_RECLAIM flag, which
doesn't really seem to be necessary. The laundromat jobs are always
kicked off via a timer, and not from memory reclaim paths. There's no
need for a rescuer thread.

Signed-off-by: Jeff Layton <jeff.layton@primarydata.com>
Signed-off-by: J. Bruce Fields <bfields@redhat.com>